third_party/blink/web_tests/external/wpt/css/css-ui/accent-color-computed-expected.txt

This is a testharness.js-based test.
PASS Property accent-color value 'initial'
PASS Property accent-color value 'inherit'
PASS Property accent-color value 'red'
PASS Property accent-color value 'blue'
PASS Property accent-color value 'auto'
FAIL Property accent-color value 'currentcolor' assert_equals: expected "currentcolor" but got "rgb(0, 0, 0)"
PASS Property accent-color value '#fff'
Harness: the test ran to completion.
